Renamed sidecarcontainer to sidecar
